working insertion electrode builder
